Update Ansible playbook and add Docker configuration
This commit makes meaningful changes in the Ansible playbook for better user directory management and font download. The play now also includes fetching the repository 'config-arch' and copying a file from it. Additionally, this adds Docker configuration for testing Ansible playbook locally with Docker, documented steps in the update README file for ease of use.