Switch base image from Arch Linux to Ubuntu in Dockerfile
In the Ansible Dockerfile, the base image has been switched from Arch Linux to Ubuntu, leading to modifications in installation commands. This alteration aims to provide better support and compatibility. The necessary modifications have been made to Ubuntu specific package management commands for the installation of Ansible and Ansible core.