Add WireGuard setup tasks in Ansible
Setup tasks for WireGuard have been defined in a new Ansible YAML file. These tasks include cloning the git repository, decrypting and copying files, and making a script executable. Furthermore, an import of these tasks was also included in the main Ansible script, along with the definition of the user and directory variables, and the installation of WireGuard and Git.